Add jwt unit tests

+++ internal/handlers/auth/jwt/logout/logout.go
@@ -19,17 +19,24 @@ func New(log *slog.Logger, tokenManager auth.JwtManager) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		log.Info("JWT  Logout user")
 
-		userID, err := strconv.Atoi(r.Context().Value("user_id").(string))
+		userIDValue := r.Context().Value("user_id")
+		if userIDValue == nil {
+			log.Error("user_id not found in context")
+			utils.SendError(w, "Invalid user context")
+			return
+		}
+
+		userID, err := strconv.Atoi(userIDValue.(string))
 		if err != nil {
-			log.Error("Invalid user ID")
-			utils.SendError(w, err.Error())
+			log.Error("Invalid user ID format")
+			utils.SendError(w, "Invalid user ID")
 			return
 		}
 
 		err = tokenManager.DeleteRefreshToken(userID)
 		if err != nil {
 			log.Error("Error deleting refresh token")
-			utils.SendError(w, err.Error())
+			utils.SendError(w, "Error deleting refresh token")
 			return
 		}
